+++ mod/src/test/java/org/shsts/tinactory/unit/logistics/CraftPortChannelTest.java
@@ -82,7 +82,7 @@ class CraftPortChannelTest {
         }
 
         @Override
-        public int getStorageAmount(TestStack stack) {
+        public long getStorageAmount(TestStack stack) {
             return 0;
         }
