+++ mod/src/test/java/org/shsts/tinactory/unit/logistics/PortFilterApiTest.java
@@ -64,7 +64,7 @@ class PortFilterApiTest {
         }
 
         @Override
-        public int getStorageAmount(TestStack stack) {
+        public long getStorageAmount(TestStack stack) {
             return 0;
         }
 
@@ -116,7 +116,7 @@ class PortFilterApiTest {
         }
 
         @Override
-        public int getStorageAmount(TestStack stack) {
+        public long getStorageAmount(TestStack stack) {
             return 0;
         }
